Split company transformer tests to seperate files
